% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{appendList}
\alias{appendList}
\title{Copy of \code{smacof:::appendList} since not exported from \code{smacof}}
\usage{
appendList(x, a)
}
\description{
Copy of \code{smacof:::appendList} since not exported from \code{smacof}
}
\keyword{internal}
