% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{bmat}
\alias{bmat}
\title{Copy of \code{smacof:::bmat} since not exported from \code{smacof}}
\usage{
bmat(diss, wgths, d, eps = 1e-12)
}
\description{
Copy of \code{smacof:::bmat} since not exported from \code{smacof}
}
\keyword{internal}
