% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmc.R
\name{check_consensus_embeddings}
\alias{check_consensus_embeddings}
\title{Check consensus embeddings for uniqueness of points}
\usage{
check_consensus_embeddings(y, fit = NULL)
}
\arguments{
\item{y}{Raw data/embeddings in new (dimension-reduced) space.}

\item{fit}{(Optional) Fitted \code{locoMDS} object corresponding to the
embeddings in \code{y}. Only used for warning messages about uniqueness of
points.}
}
\value{
Original consensus embeddings if sufficiently unique; otherwise,
the rounded consensus embeddings with a warning.
}
\description{
Check consensus embeddings for uniqueness of points
}
\keyword{internal}
