% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{checkdiss}
\alias{checkdiss}
\title{Copy of \code{smacof:::checkdiss} since not exported from \code{smacof}}
\usage{
checkdiss(diss)
}
\description{
Copy of \code{smacof:::checkdiss} since not exported from \code{smacof}
}
\keyword{internal}
