% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMDS.R
\name{coMDS}
\alias{coMDS}
\title{Consensus Multidimensional Scaling (CoMDS)}
\usage{
coMDS(
  D_list,
  embed_list = NULL,
  ndim = 2,
  constraint = c("indscal", "idioscal", "identity"),
  ...
)
}
\arguments{
\item{D_list}{List of distance matrices.}

\item{embed_list}{Optional list of embeddings to compute distances from. If
provided, the pairwise Euclidean distances will be computed between the
embedded samples and used in lieu of \code{D_list}.}

\item{ndim}{Number of dimensions for the output embedding. Default is 2.}

\item{constraint}{Type of weight constraint to use in the CoMDS algorithm.
Options are "indscal" (default), "idioscal", or "identity"}

\item{...}{Additional arguments passed to the \code{\link[=smacofIndDiff]{smacofIndDiff()}} function.}
}
\value{
A list containing the following components:
\itemize{
\item \code{gspace}: Consensus embedding matrix of size \verb{n x ndim}, where \code{n} is the
number of samples.
\item \code{cweights}: A list of learned weight matrices for each input embedding/
distance matrix.
\item \code{stress}: The final stress value of the CoMDS solution.
\item \code{spp}: Stress per point (in percent).
\item \code{rss}: Residual sum of squares.
\item \code{spps}: Stress per point per input embedding/distance matrix (in percent,
conditional on the input embedding/distance matrix).
\item \code{sps}: Stress per input embedding/distance matrix (in percent).
\item \code{sps_norm}: Stress per input embedding/distance matrix (in percent),
normalized by sum of source weights.
\item \code{ndim}: Number of dimensions in consensus embedding.
\item \code{model}: Type of smacof model.
\item \code{niter}: Number of iterations run.
\item \code{nobj}: Number of points/samples.
\item \code{type}: MDS type used (e.g., "interval", "ratio", "ordinal", or "mspline").
\item \code{constraint}: Type of weight constraint used in the CoMDS algorithm.
\item \code{call}: The matched call to the function.
}
}
\description{
Consensus Multidimensional Scaling (CoMDS)
}
\examples{
data(iris)
# remove duplicates so that tSNE can run
X <- dplyr::distinct(iris[, 1:4])

# fit various dimension reduction methods
pca_scores <- prcomp(X, center = TRUE, scale = TRUE)$x
tsne_scores <- Rtsne::Rtsne(X, dims = 2, perplexity = 30, verbose = FALSE)$Y
umap_scores <- umap::umap(X, n_components = 2, verbose = FALSE)$layout
dr_list <- list(
  pca = pca_scores,
  tsne = tsne_scores,
  umap = umap_scores
)

# fit coMDS using dimension reduction embeddings directly as input
comds_out <- coMDS(embed_list = dr_list, ndim = 2)

# fit coMDS using distance matrices as input
dist_list <- purrr::map(dr_list, ~ dist(.x, method = "euclidean"))
comds_out <- coMDS(D_list = dist_list, ndim = 2)

# can re-run with larger number of iterations or smaller tolerance if desired
comds_out <- coMDS(D_list = dist_list, ndim = 2, itmax = 500, eps = 1e-8)

}
