% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-locoMDS.R
\name{get_ts}
\alias{get_ts}
\title{Get t values for regularization in LoCoMDS}
\usage{
get_ts(delta, weights, tau, percentile)
}
\description{
Get t values for regularization in LoCoMDS
}
\keyword{internal}
