% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-locoMDS.R
\name{init_local_weights}
\alias{init_local_weights}
\title{Get weight matrices based upon local neighborhoods}
\usage{
init_local_weights(neighbor_list)
}
\description{
Get weight matrices based upon local neighborhoods
}
\keyword{internal}
