% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{initConf}
\alias{initConf}
\title{Copy of \code{smacof:::initConf} since not exported from \code{smacof}}
\usage{
initConf(init, diss, n, p, inddiff = FALSE)
}
\description{
Copy of \code{smacof:::initConf} since not exported from \code{smacof}
}
\keyword{internal}
