% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{initWeights}
\alias{initWeights}
\title{Copy of \code{smacof:::initWeights} since not exported from \code{smacof}}
\usage{
initWeights(diss)
}
\description{
Copy of \code{smacof:::initWeights} since not exported from \code{smacof}
}
\keyword{internal}
