% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmc.R
\name{lcmc}
\alias{lcmc}
\title{Evaluate Local Continuity Meta-Criterion (LCMC)}
\usage{
lcmc(x, y, ks, fit = NULL)
}
\arguments{
\item{x}{Raw data/embeddings in original space.}

\item{y}{Raw data/embeddings in new (dimension-reduced) space.}

\item{ks}{Vector of neighborhood sizes to evaluate LCMC.}

\item{fit}{(Optional) Fitted \code{locoMDS} object corresponding to the
embeddings in \code{y}. Only used for warning messages about uniqueness of
points.}
}
\value{
A data frame with columns:
\itemize{
\item \code{k}: Neighborhood size.
\item \code{lcmc}: Local Continuity Meta-Criterion value for the given neighborhood
size.
\item \code{adjusted_lcmc}: Adjusted LCMC value, accounting for the neighborhood size.
}
}
\description{
Evaluate Local Continuity Meta-Criterion (LCMC)
}
\examples{
# generate example data
n <- 100
p1 <- 4
p2 <- 2
x <- matrix(rnorm(n * p1), nrow = n, ncol = p1)
y <- matrix(rnorm(n * p2), nrow = n, ncol = p2)

# evaluate LCMC for neighborhood sizes 5, 10, and 20
lcmc_out <- lcmc(x, y, ks = c(5, 10, 20))

}
