% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{normDissN}
\alias{normDissN}
\title{Normalize dissimilarities for CoMDS, allowing different sources to have
different overall weights}
\usage{
normDissN(diss, wghts, m)
}
\description{
Normalize dissimilarities for CoMDS, allowing different sources to have
different overall weights
}
\keyword{internal}
