% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{repList}
\alias{repList}
\title{Copy of \code{smacof:::repList} since not exported from \code{smacof}}
\usage{
repList(x, n)
}
\description{
Copy of \code{smacof:::repList} since not exported from \code{smacof}
}
\keyword{internal}
