% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smacofIndDiff.R
\name{smacofIndDiff}
\alias{smacofIndDiff}
\title{SMACOF for Individual Differences}
\usage{
smacofIndDiff(
  delta,
  ndim = 2,
  type = c("ratio", "interval", "ordinal", "mspline"),
  constraint = c("indscal", "idioscal", "identity"),
  weights = NULL,
  weightmat = NULL,
  weightlam = 0,
  init = "torgerson",
  ties = "primary",
  verbose = FALSE,
  modulus = 1,
  itmax = 100,
  eps = 1e-06,
  spline.degree = 2,
  spline.intKnots = 2
)
}
\arguments{
\item{delta}{A list of dissimilarity matrices or a list objects of class \code{dist}}

\item{ndim}{Number of dimensions}

\item{type}{MDS type: \code{"interval"}, \code{"ratio"}, \code{"ordinal"} (nonmetric MDS), or \code{"mspline"}}

\item{constraint}{Either \code{"indscal"}, \code{"idioscal"}, or \code{"identity"} (see details)}

\item{weights}{Vector of length equal to the number of distance matrices.
If provided, weights for each distance matrix are multiplied by the
corresponding value in \code{weights}. Default is \code{NULL} which will give
each distance matrix the same weight (i.e., scaled by 1).}

\item{weightmat}{Optional matrix with dissimilarity weights}

\item{weightlam}{Regularization parameter for weight matrices to avoid
singularity matrices. Default is 0 (i.e., no regularization). If
regularization is needed, set to a small value (e.g., 1e-6).}

\item{init}{Matrix with starting values for configurations (optional)}

\item{ties}{Tie specification for non-metric MDS}

\item{verbose}{If \code{TRUE}, intermediate stress is printed out}

\item{modulus}{Number of smacof iterations per monotone regression call}

\item{itmax}{Maximum number of iterations}

\item{eps}{Convergence criterion}

\item{spline.degree}{Degree of the spline for \code{"mspline"} MDS type}

\item{spline.intKnots}{Number of interior knots of the spline for \code{"mspline"} MDS type}
}
\description{
Variant of \code{smacof::smacofIndDiff()} but with regularization
(\code{weightlam}) in case of singularity in weight matrices (e.g., due to
missingness) and optional weight scaling (\code{weights}) which multiples
weights for each data source by some amount.
}
