% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{spp}
\alias{spp}
\title{Copy of \code{smacof:::spp} since not exported from \code{smacof}}
\usage{
spp(dhat, confdiss, wgths)
}
\description{
Copy of \code{smacof:::spp} since not exported from \code{smacof}
}
\keyword{internal}
