% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{strucprep}
\alias{strucprep}
\title{Copy of \code{smacof:::strucprep} since not exported from \code{smacof}}
\usage{
strucprep(x)
}
\description{
Copy of \code{smacof:::strucprep} since not exported from \code{smacof}
}
\keyword{internal}
