% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-smacof.R
\name{vmat}
\alias{vmat}
\title{Regularized version of \code{smacof::vmat()} to avoid singularity issues}
\usage{
vmat(wgths, weightlam = 0)
}
\description{
Regularized version of \code{smacof::vmat()} to avoid singularity issues
}
\keyword{internal}
